[Prover] Consolidate prove() logic before refactoring.
