+++ prover-service/src/lib.rs
@@ -10,11 +10,9 @@ pub mod handlers;
 pub mod input_processing;
 pub mod jwk_fetching;
 pub mod keyless_config;
-pub mod load_vk;
 pub mod metrics;
 pub mod prover_config;
 pub mod prover_state;
-pub mod proving;
 pub mod training_wheels;
 pub mod utils;
 pub mod witness_gen;
